HOLO-208 <refactor: 함수 선언 위치 변경>
